+++ src/main/java/teamkiim/koffeechat/domain/email/controller/dto/AuthCodeCheckRequest.java
@@ -1,5 +1,6 @@
 package teamkiim.koffeechat.domain.email.controller.dto;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,7 +11,9 @@ import teamkiim.koffeechat.domain.email.dto.request.AuthCodeCheckServiceRequest;
 @AllArgsConstructor
 public class AuthCodeCheckRequest {
 
+    @NotBlank(message = "이메일을 입력해주세요")
     private String email;
+    @NotBlank(message = "인증 코드를 입력해주세요")
     private String code;
 
     public AuthCodeCheckServiceRequest toServiceRequest(){
