+++ src/main/java/teamkiim/koffeechat/domain/member/repository/MemberRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
 
-    public Optional<Member> findById(Long id);
+    Optional<Member> findById(Long id);
 
-    public Optional<Member> findByEmail(String email);
+    Optional<Member> findByEmail(String email);
 }
