+++ src/main/java/teamkiim/koffeechat/domain/post/community/dto/request/ModifyCommunityPostServiceRequest.java
@@ -1,21 +1,17 @@
-package teamkiim.koffeechat.domain.post.community.service.dto.request;
+package teamkiim.koffeechat.domain.post.community.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-import java.util.List;
-
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class SaveCommunityPostServiceRequest {
+public class ModifyCommunityPostServiceRequest {
 
     private Long id;
     private String title;
     private String bodyContent;
-    private List<Long> fileIdList;
 }
