+++ src/main/java/teamkiim/koffeechat/domain/post/community/dto/request/SaveCommunityPostServiceRequest.java
@@ -1,17 +1,21 @@
-package teamkiim.koffeechat.domain.post.community.service.dto.request;
+package teamkiim.koffeechat.domain.post.community.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDateTime;
+import java.util.List;
+
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class ModifyCommunityPostServiceRequest {
+public class SaveCommunityPostServiceRequest {
 
     private Long id;
     private String title;
     private String bodyContent;
+    private List<Long> fileIdList;
 }
