+++ src/main/java/teamkiim/koffeechat/domain/post/community/dto/response/VoteItemInfoDto.java
@@ -1,4 +1,4 @@
-package teamkiim.koffeechat.domain.vote.service.dto;
+package teamkiim.koffeechat.domain.post.community.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -10,14 +10,14 @@ import teamkiim.koffeechat.domain.vote.domain.VoteItem;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class SaveVoteRecordServiceDto {
+public class VoteItemInfoDto {
 
-    private Long id;                                        // 투표 항목 id
+    private Long id;
     private String itemText;
     private Long voteCount;
 
-    public static SaveVoteRecordServiceDto of(VoteItem voteItem) {
-        return SaveVoteRecordServiceDto.builder()
+    public static VoteItemInfoDto of(VoteItem voteItem) {
+        return VoteItemInfoDto.builder()
                 .id(voteItem.getId())
                 .itemText(voteItem.getItemText())
                 .voteCount(voteItem.getVoteCount())
