+++ src/main/java/teamkiim/koffeechat/domain/post/community/service/CommunityPostService.java
@@ -15,19 +15,18 @@ import teamkiim.koffeechat.domain.member.domain.Member;
 import teamkiim.koffeechat.domain.member.repository.MemberRepository;
 import teamkiim.koffeechat.domain.post.community.domain.CommunityPost;
 import teamkiim.koffeechat.domain.post.community.repository.CommunityPostRepository;
-import teamkiim.koffeechat.domain.post.community.service.dto.request.ModifyCommunityPostServiceRequest;
-import teamkiim.koffeechat.domain.post.community.service.dto.request.SaveCommunityPostServiceRequest;
-import teamkiim.koffeechat.domain.post.community.service.dto.response.CommentInfoDto;
-import teamkiim.koffeechat.domain.post.community.service.dto.response.CommunityPostListResponse;
-import teamkiim.koffeechat.domain.post.community.service.dto.response.CommunityPostResponse;
-import teamkiim.koffeechat.domain.post.community.service.dto.response.VoteResponse;
-import teamkiim.koffeechat.domain.postlike.domain.PostLike;
+import teamkiim.koffeechat.domain.post.community.dto.request.ModifyCommunityPostServiceRequest;
+import teamkiim.koffeechat.domain.post.community.dto.request.SaveCommunityPostServiceRequest;
+import teamkiim.koffeechat.domain.post.community.dto.response.CommentInfoDto;
+import teamkiim.koffeechat.domain.post.community.dto.response.CommunityPostListResponse;
+import teamkiim.koffeechat.domain.post.community.dto.response.CommunityPostResponse;
+import teamkiim.koffeechat.domain.post.community.dto.response.VoteResponse;
 import teamkiim.koffeechat.domain.postlike.repository.PostLikeRepository;
 import teamkiim.koffeechat.domain.vote.domain.Vote;
 import teamkiim.koffeechat.domain.vote.repository.VoteRepository;
 import teamkiim.koffeechat.domain.vote.service.VoteService;
-import teamkiim.koffeechat.domain.vote.service.dto.request.ModifyVoteServiceRequest;
-import teamkiim.koffeechat.domain.vote.service.dto.request.SaveVoteServiceRequest;
+import teamkiim.koffeechat.domain.vote.dto.request.ModifyVoteServiceRequest;
+import teamkiim.koffeechat.domain.vote.dto.request.SaveVoteServiceRequest;
 
 import java.util.ArrayList;
 import java.util.List;
