+++ src/main/java/teamkiim/koffeechat/domain/vote/service/VoteService.java
@@ -11,15 +11,15 @@ import teamkiim.koffeechat.domain.member.domain.Member;
 import teamkiim.koffeechat.domain.member.repository.MemberRepository;
 import teamkiim.koffeechat.domain.post.common.domain.Post;
 import teamkiim.koffeechat.domain.post.common.repository.PostRepository;
-import teamkiim.koffeechat.domain.vote.controller.dto.request.SaveVoteRecordRequest;
+import teamkiim.koffeechat.domain.vote.controller.dto.SaveVoteRecordRequest;
 import teamkiim.koffeechat.domain.vote.domain.Vote;
 import teamkiim.koffeechat.domain.vote.domain.VoteItem;
 import teamkiim.koffeechat.domain.vote.domain.VoteRecord;
 import teamkiim.koffeechat.domain.vote.repository.VoteRecordRepository;
 import teamkiim.koffeechat.domain.vote.repository.VoteRepository;
-import teamkiim.koffeechat.domain.vote.service.dto.SaveVoteRecordServiceDto;
-import teamkiim.koffeechat.domain.vote.service.dto.request.ModifyVoteServiceRequest;
-import teamkiim.koffeechat.domain.vote.service.dto.request.SaveVoteServiceRequest;
+import teamkiim.koffeechat.domain.vote.dto.SaveVoteRecordServiceDto;
+import teamkiim.koffeechat.domain.vote.dto.request.ModifyVoteServiceRequest;
+import teamkiim.koffeechat.domain.vote.dto.request.SaveVoteServiceRequest;
 
 import java.util.List;
 import java.util.stream.Collectors;
