+++ src/test/java/teamkiim/koffeechat/chat/service/ChatRoomServiceTest.java
@@ -58,8 +58,8 @@ class ChatRoomServiceTest extends TestSupport {
         Long chatRoomId = (Long) responseEntity.getBody();
         Optional<ChatRoom> savedChatRoom = chatRoomRepository.findById(chatRoomId);
         assertThat(savedChatRoom).isNotNull();
-        assertThat(savedChatRoom.get().getCreateMember().getId()).isEqualTo(memberId);
-        assertThat(savedChatRoom.get().getPost().getId()).isEqualTo(postId);
+//        assertThat(savedChatRoom.get().getCreateMember().getId()).isEqualTo(memberId);
+//        assertThat(savedChatRoom.get().getPost().getId()).isEqualTo(postId);
 
     }
 
@@ -80,8 +80,8 @@ class ChatRoomServiceTest extends TestSupport {
                 .bodyContent("Sample body content")
                 .viewCount(0L)
                 .likeCount(0L)
-                .createdTime(LocalDateTime.of(2024, 6, 12, 10, 0))
-                .modifiedTime(null)
+//                .createdTime(LocalDateTime.of(2024, 6, 12, 10, 0))
+//                .modifiedTime(null)
                 .isEditing(false)
                 .skillCategoryList(null)
                 .build();
